Set up basic Hubspot API client boilerplate, start refactoring Slack modules into single context
