+++ lib/chat_api/customers/customer.ex
@@ -47,11 +47,8 @@ defmodule ChatApi.Customers.Customer do
     timestamps()
   end
 
-  @spec changeset(
-          {map, map} | %{:__struct__ => atom | %{__changeset__: map}, optional(atom) => any},
-          :invalid | %{optional(:__struct__) => none, optional(atom | binary) => any}
-        ) :: Ecto.Changeset.t()
   @doc false
+  @spec changeset(any(), map()) :: Ecto.Changeset.t()
   def changeset(customer, attrs) do
     customer
     |> cast(attrs, [
